moved splitArr function
